Add files via upload
1.real-time hand display
This code opens the camera and tracks one hand in real time using MediaPipe. It draws green points on the hand joints and white lines connecting them. The video with these markings is displayed on the screen, and the program stops when the q key is pressed.
2.draw finger coordinates
This code tracks a single finger (index finger) of a hand in real time using MediaPipe. A blue circle is drawn on the fingertip, and the X and Y coordinates in pixels are displayed next to it. The video also shows the hand landmarks connected by lines, and the program stops when the q key is pressed.